requirements.txt
python-telegram-bot>=20.0
python-dotenv>=0.19.0
Flask>=2.0.0